+++ data_structures/interpolation_base.pyx
@@ -1,7 +1,7 @@
 cdef class InterpolationBase:
-    def __call__(self, float x):
+    def __call__(self, double x):
         x = min(max(x, 0), 1)
         return self.evaluate(x)
 
-    cdef float evaluate(self, float x):
+    cdef double evaluate(self, double x):
         raise NotImplementedError()
